Time: 18 ms (63.54%), Space: 44.1 MB (33.81%) - LeetHub

+++ 0018-4sum/0018-4sum.java
@@ -22,7 +22,7 @@ class Solution {
                         while(k<l && nums[k] == nums[k-1])  k++;
                         while(k<l && nums[l] == nums[l+1])  l--;
                     }
-                    else if(sum<(long)target) k++;
+                    else if(sum<target) k++;
                     else    l--;
                 }
             }
